requirements.txt
gTTS==2.4.0
python-docx==1.1.2
pdfplumber==0.11.0
tqdm==4.66.2
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
